+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 0.12.9
+
+### Patch Changes
+
+-   Updated dependencies [b6d0c67]
+    -   @sap-ux/odata-service-writer@0.14.6
+    -   @sap-ux/ui5-config@0.14.5
+    -   @sap-ux/ui5-application-writer@0.14.6
+
 ## 0.12.8
 
 ### Patch Changes